==> 06_02_vectorized_backtest (3).ipynb <==
# Vectorized Backtest

import warnings
warnings.filterwarnings('ignore')

from pathlib import Path
from time import time
import datetime

import numpy as np
import pandas as pd
import pandas_datareader.data as web

from scipy.stats import spearmanr

import matplotlib.pyplot as plt
from matplotlib.ticker import FuncFormatter
import seaborn as sns

sns.set_style('whitegrid')
np.random.seed(42)

# settings iniciales

#periodo target. 1 implica que queremos predecir 1 periodo siguiente. (una semana en caso de datos semanales)
periodo_target=1 

## Load Data

### Return Predictions



# Lee el archivo hdf
store = pd.HDFStore('../data/predictions.h5')

# Obtiene las claves
keys = store.keys()

# Cierra el archivo hdf
store.close()

# Imprime las claves
print(keys)

#tomamos los resultados de las 10 mejores configuraciones de los predictores
predictions = pd.concat([pd.read_hdf('../data/predictions.h5', 'lgb/train/01'),
                   pd.read_hdf('../data/predictions.h5', 'lgb/test/01').drop('y_test', axis=1)])
#tomamos la media
predictions = (predictions.loc[~predictions.index.duplicated()]#si se duplican indices tomara los del train, ose los primeros
                   .iloc[:, :10]
                   .mean(1)
                   .sort_index()
                   .dropna()
                  .to_frame('prediction'))
tickers = predictions.index.get_level_values('ticker').unique().tolist()#modificado ticker por symbol


datos_train=(pd.read_hdf('../data/predictions.h5', 'lgb/train/01'))
datos_test=pd.read_hdf('../data/predictions.h5', 'lgb/test/01')

#fin del train
datos_train.sort_index().loc['XLY'].tail(1)

#comienzo del test
datos_test.sort_index().loc['XLY'].head(1)

predicciones=(predictions
            .unstack('ticker')
            .prediction)


predicciones.info()

### SP500 Benchmark

#sp500 = web.DataReader('SP500', 'fred', '2014', '2018').pct_change()

#sp500.info()

## Compute Forward Returns

DATA_STORE = '../data/assets.h5'

idx = pd.IndexSlice

with pd.HDFStore(DATA_STORE) as store:
    data = (store['data_raw']
            .sort_index()
            .loc[idx[:, :'2024'], :])

data.info()

#localizamos target_1m para las fechas y tickers de predictions
fwd_returns = (data.loc[idx[predictions.index.get_level_values('ticker').unique(),
               predictions.index.get_level_values('date').unique()],:]
               .sort_index(ascending=False))

fwd_returns

# Definimos period target

#creamos la variable target_{lookahead} mesesm
fwd_returns[f'target_{periodo_target}m'] = data.groupby(level='ticker')[f'return_{periodo_target}m'].shift(-periodo_target)

#se hace la asignación del retorno de un perdio fwd

fwd_returns = fwd_returns[f'target_{periodo_target}m'].unstack('ticker').sort_index()

#eliminamos SPY
#fwd_returns.drop('SPY.US', axis=1, inplace=True)
fwd_returns.head()


media = fwd_returns.mean(axis=1).shift(1)
media.name = 'mediasec'

## Generate Signals

#eliminamos SPY si comparamos con benchmark
#predicciones.drop('SPY.US', axis=1, inplace=True)
predicciones

N_LONG = N_SHORT = 4

long_signals = ((predicciones
                .where(predicciones > 0)
                .rank(axis=1, ascending=False, method='first') < N_LONG)
                .astype(int))
short_signals = ((predicciones
                  .where(predicciones < 0)
                  .rank(axis=1, method='first') < N_SHORT)
                 .astype(int))

## Compute Portfolio Returns

long_returns = long_signals.mul(fwd_returns).replace(0, np.nan).mean(axis=1).fillna(0)
short_returns = short_signals.mul(-fwd_returns).replace(0, np.nan).mean(axis=1).fillna(0)
strategy = long_returns.add(short_returns).to_frame('Strategy')

short_signals.loc['2019':].describe()

short_signals.loc['2019':].sum()

long_signals.loc['2019':].describe()

long_signals.loc['2019':].sum()

## Plot results

fig, axes = plt.subplots(ncols=2, figsize=(14,5))
strategy.loc['2019':].join(media).add(1).cumprod().sub(1).plot(ax=axes[0], title='Cumulative Return')
#strategy.join(sp500).add(1).cumprod().sub(1).plot(ax=axes[0], title='Cumulative Return')
sns.distplot(strategy.dropna(), ax=axes[1], hist=False, label='Strategy')
sns.distplot(media, ax=axes[1], hist=False, label='media')
axes[1].set_title('Daily Standard Deviation')
axes[0].yaxis.set_major_formatter(FuncFormatter(lambda y, _: '{:.0%}'.format(y)))
axes[1].xaxis.set_major_formatter(FuncFormatter(lambda y, _: '{:.0%}'.format(y)))
sns.despine()
fig.tight_layout();

#en vez de s&P, la media
fig, axes = plt.subplots(ncols=2, figsize=(14,5))
strategy.loc['2019':].join(media).add(1).cumprod().sub(1).plot(ax=axes[0], title='Cumulative Return')
sns.distplot(strategy.dropna(), ax=axes[1], hist=False, label='Strategy')
sns.distplot(media, ax=axes[1], hist=False, label='media')
axes[1].set_title('Daily Standard Deviation')
axes[0].yaxis.set_major_formatter(FuncFormatter(lambda y, _: '{:.0%}'.format(y)))
axes[1].xaxis.set_major_formatter(FuncFormatter(lambda y, _: '{:.0%}'.format(y)))
sns.despine()
fig.tight_layout();

res = strategy.join(media).dropna()

res.std()

fwd_returns.loc['2019':].cumsum().plot(figsize=(14,5))

solo largos

#sólo largos
strategy_long = long_returns.to_frame('Strategy')

fig, axes = plt.subplots(ncols=2, figsize=(14,5))
strategy_long.loc['2019':].join(media).add(1).cumprod().sub(1).plot(ax=axes[0], title='Cumulative Return')
sns.distplot(strategy_long.dropna(), ax=axes[1], hist=False, label='Strategy')
sns.distplot(media, ax=axes[1], hist=False, label='media')
axes[1].set_title('Daily Standard Deviation')
axes[0].yaxis.set_major_formatter(FuncFormatter(lambda y, _: '{:.0%}'.format(y)))
axes[1].xaxis.set_major_formatter(FuncFormatter(lambda y, _: '{:.0%}'.format(y)))
sns.despine()
fig.tight_layout();

long_signals.loc[:'2019'].sum()

solo cortos

#sólo CORTOS
strategy_short = short_returns.to_frame('Strategy')

fig, axes = plt.subplots(ncols=2, figsize=(14,5))
strategy_short.loc['2019':].join(media).add(1).cumprod().sub(1).plot(ax=axes[0], title='Cumulative Return')
sns.distplot(strategy_short.dropna(), ax=axes[1], hist=False, label='Strategy')
sns.distplot(media, ax=axes[1], hist=False, label='media')
axes[1].set_title('Daily Standard Deviation')
axes[0].yaxis.set_major_formatter(FuncFormatter(lambda y, _: '{:.0%}'.format(y)))
axes[1].xaxis.set_major_formatter(FuncFormatter(lambda y, _: '{:.0%}'.format(y)))
sns.despine()
fig.tight_layout();
